fix: 支持 SCHEDULE_FETCH_MARKETS_INTERVAL_HOURS 小数输入
将类型从 int 改为 float，支持 0.5 小时（30分钟）等设置

via [Happy](https://happy.engineering)

+++ src/config.py
@@ -455,7 +455,7 @@ class TaskScheduleSettings(BaseSettings):
         extra="ignore",
     )
 
-    fetch_markets_interval_hours: int = Field(
+    fetch_markets_interval_hours: float = Field(
         default=2,
         alias="SCHEDULE_FETCH_MARKETS_INTERVAL_HOURS",
         gt=0,
